REPO-2025 Main Bamboo TestSuites now exclude DBTests and NonBuildTests
Required a rework of the way the test classes were specied. Needed to use
@Suite.SuiteClasses. As a result the AllRepositoryTestsCatalogue was removed.
Also removed TestSuites it pulled in. They turned out to be of limited use
for developers. Thewre are still a number of TestSuites not run by Bamboo.

Pulled in a new version of alfresco-core which had the NonBuildTests interface.