REPO-3972 doc to pdf is not sent to the Transform Service
The RemoteTransformClient was not sending the transform to the Transform Service because, the "pdf" rendition had an extra allowEnlargement option, which had been added in 6.1 EA3 as a way to force the conversion to use the ImageTransformationOptions sub class rather then TransformationOptions. While investigation this bug, it was found the type should have been SWFTransformationOptions, which had been incorrectly added in ACS 5.2.