SEARCH-1351 Add unit tests for SolrSQLHttpClient.
This involved changing the return type of postSolrQuery. The cast
taking place in this file did not make sense to me, since the two
types are not in the same type hierarchy. However it is obviously
working at the moment, so I have simply moved the location of the
cast. I couldn't find any references to the protected method anyway,
so it did not seem risky to update.