REPO-4613: add aspect to a private working copy
Using CMIS 1.1, it's not possible to add an aspect or update the properties for a private copy.
Once the checkout is done, the aspect P:cm:checkedOut is added by the system and is present in the pwc.
Ignoring it during the properties update as it's done for the aspects added by the system, solve the issue.