MNT-18895: Disabling/Enabling behaviour enhancement
This an extension of the BehaviourFilter service for smart folders.
In case of virtual nodes, the disbaling/enabling service must do the requests for the actual nodes otherwise it is only done for the virtual ones
Therefore, the actual nodes get updated since the hadn't been touched by the BehaviourFilter service.