MNT-18895: Extension of Behaviour Filter for Smart Folders
The MNT-18895 ticket has highlighted the fact that behaviour filter service did not have an extension when it comes to smart folders.

The issue had to do with the creation of renditions as children of documents, when a document hasn't got any rendition yet, and the rendition is created as a new node from a smart folder, it is only the virtual node that gets its aspect cm:auditable disbaled and not the actual node linked to it.
The cm:auditable is responsible for certain updates of node's properties, in particular the properties cm:modifier and cm:modified.

This merge delivers an extension for behaviour filter service as it is the case already for other services (node, coci, lock, fileFolder..)

* Extension of the BehaviourFilter service for smart folders:
   - VirtualBehaviourFilterExtension: specific implementation of BehaviourFilter service, executed in smart folder context
   - BehaviourFilterExtension/BehaviourFilterTrait/annotations on top of override BehaviourFilterImpl methods: allow AOP mechanism
   - smartfolder-context: declaration of the new beans and injection into the virtualisation bundle
   - VirtualBehaviourFilterExtensionTest and AppContextExtraTestSuite: to ensure that the extension is used instead of the original implementation

(cherry picked from commit 9399a743fdae9d3be5fe688e7cca8a1249701d0a)